# packages/contracts/openapi.yaml
openapi: 3.1.0
...
info:
  title: Project Booth API
  summary: Platform-neutral API contract for Project Booth clients.
  description: |
    Product endpoints use path versioning beginning with `/v1`. Operational
    liveness remains at the unversioned `/health/live` path so container
    orchestration does not depend on a product API version.

    Identity endpoints exchange provider credentials for Project Booth
    sessions. Game endpoints trust only the resulting internal account and
    never a platform-provider identifier.
  version: 0.3.0
...
servers:
  - url: "{scheme}://{authority}"
    description: Environment-selected API origin.
    variables:
      scheme:
        default: https
        enum:
          - https
          - http
        description: HTTPS in deployed environments; HTTP is local-only.
      authority:
        default: api.invalid
        description: Host and optional port supplied by the client environment.
...
security:
  - bearerAuth: []
paths:
...
  /v1/auth/exchange:
    post:
      tags:
        - Identity
      summary: Exchange an identity-provider credential
      operationId: exchangeIdentityCredential
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExchangeCredentialRequest"
      responses:
        "201":
          description: The credential was verified and a session was created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeCredentialResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/auth/refresh:
    post:
      tags:
        - Identity
      summary: Rotate a refresh token
      operationId: refreshIdentitySession
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshSessionRequest"
      responses:
        "200":
          description: A fresh access and refresh token pair.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/auth/logout:
    post:
      tags:
        - Identity
      summary: Revoke the current session
      operationId: logoutIdentitySession
      responses:
        "204":
          description: The current session was revoked.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/account:
    get:
      tags:
        - Account
      summary: Get the current private account projection
      operationId: getCurrentAccount
      responses:
        "200":
          description: Current account state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    delete:
      tags:
        - Account
      summary: Request account deletion
      description: |
        Immediately revokes every account session and queues asynchronous
        identity and profile cleanup. A retained internal tombstone preserves
        referential integrity for match and safety history.
      operationId: requestAccountDeletion
      responses:
        "202":
          description: Deletion is pending or was already completed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountDeletionAccepted"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/profile:
    patch:
      tags:
        - Profile
      summary: Update the current pseudonymous profile
      operationId: updateCurrentProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated public profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicProfile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/profiles/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        description: Internal Project Booth player identifier.
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Profile
      summary: Get a match-safe public profile
      operationId: getPublicProfile
      responses:
        "200":
          description: Match-safe pseudonymous profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/matchmaking/tickets:
    post:
      tags:
        - Matchmaking
      summary: Create or recover the account's active matchmaking ticket
      operationId: createMatchmakingTicket
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMatchmakingTicketRequest"
      responses:
        "200":
          description: The account already had an active ticket.
          headers:
            Idempotency-Replayed:
              $ref: "#/components/headers/IdempotencyReplayed"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchmakingTicket"
        "201":
          description: A durable ticket was created.
          headers:
            Idempotency-Replayed:
              $ref: "#/components/headers/IdempotencyReplayed"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchmakingTicket"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/matchmaking/tickets/{ticketId}:
    parameters:
      - $ref: "#/components/parameters/TicketId"
    get:
      tags:
        - Matchmaking
      summary: Inspect a matchmaking ticket
      operationId: getMatchmakingTicket
      responses:
        "200":
          description: Current durable ticket state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchmakingTicket"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        A short-lived Project Booth access token whose encoding is opaque to
        clients. Authentication behavior is implemented in a later milestone.
  parameters:
...
    TicketId:
      name: ticketId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        RFC 9562 UUID generated for one mutation attempt. Keys are scoped to
        the authenticated account and operation; replay returns the original
        committed response.
      schema:
        type: string
        format: uuid
  headers:
    RequestId:
      description: Request trace identifier for support and diagnostics.
      schema:
        type: string
        minLength: 1
        maxLength: 128
    IdempotencyReplayed:
      description: Whether this response replays an earlier committed result.
      schema:
        type: boolean
...
    WWWAuthenticate:
      description: Bearer authentication challenge.
      schema:
        type: string
  schemas:
...
    DeviceRegistration:
      type: object
      additionalProperties: false
      required:
        - installationId
        - platform
      properties:
        installationId:
          type: string
          format: uuid
          description: App-installation identifier, not a hardware identifier.
        platform:
          type: string
          enum:
            - ios
            - test
        appVersion:
          type: string
          minLength: 1
          maxLength: 64
    ExchangeCredentialRequest:
      type: object
      additionalProperties: false
      required:
        - credential
        - device
      properties:
        credential:
          type: string
          minLength: 1
          maxLength: 8192
          description: |
            Provider credential selected by the server environment. Production
            Apple credentials are verified against Apple's signing keys.
        nonce:
          type: string
          minLength: 1
          maxLength: 128
        device:
          $ref: "#/components/schemas/DeviceRegistration"
    RefreshSessionRequest:
      type: object
      additionalProperties: false
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          minLength: 1
          maxLength: 256
    TokenPair:
      type: object
      additionalProperties: false
      required:
        - accessToken
        - accessTokenExpiresAt
        - refreshToken
        - refreshTokenExpiresAt
        - tokenType
      properties:
        accessToken:
          type: string
          minLength: 1
          maxLength: 256
        accessTokenExpiresAt:
          type: string
          format: date-time
        refreshToken:
          type: string
          minLength: 1
          maxLength: 256
        refreshTokenExpiresAt:
          type: string
          format: date-time
        tokenType:
          type: string
          const: Bearer
    PublicProfile:
      type: object
      additionalProperties: false
      required:
        - userId
        - displayName
        - avatarKey
        - progressionLevel
      properties:
        userId:
          type: string
          format: uuid
        displayName:
          type: string
          minLength: 3
          maxLength: 24
        avatarKey:
          type: string
          minLength: 1
          maxLength: 64
        progressionLevel:
          type: integer
          minimum: 1
    Account:
      type: object
      additionalProperties: false
      required:
        - userId
        - status
        - deletionStatus
        - profile
      properties:
        userId:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - active
            - deletion_pending
            - deleted
        deletionStatus:
          type: string
          enum:
            - none
            - pending
            - completed
        profile:
          $ref: "#/components/schemas/PublicProfile"
    ExchangeCredentialResponse:
      type: object
      additionalProperties: false
      required:
        - account
        - tokens
      properties:
        account:
          $ref: "#/components/schemas/Account"
        tokens:
          $ref: "#/components/schemas/TokenPair"
    UpdateProfileRequest:
      type: object
      additionalProperties: false
      required:
        - displayName
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 128
    AccountDeletionAccepted:
      type: object
      additionalProperties: false
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - pending
            - completed
    CreateMatchmakingTicketRequest:
      type: object
      additionalProperties: false
      required:
        - compatibilityVersion
        - language
        - region
      properties:
        compatibilityVersion:
          type: integer
          minimum: 1
          maximum: 2147483647
        language:
          type: string
          pattern: ^[a-z]{2,3}(?:-[A-Z]{2})?$
        region:
          type: string
          pattern: ^[a-z0-9][a-z0-9-]{0,31}$
    MatchmakingTicket:
      type: object
      additionalProperties: false
      required:
        - compatibilityVersion
        - createdAt
        - language
        - readyConfirmed
        - region
        - rulesetId
        - rulesetVersion
        - status
        - ticketId
        - updatedAt
      properties:
        compatibilityVersion:
          type: integer
          minimum: 1
        createdAt:
          type: string
          format: date-time
        language:
          type: string
        matchId:
          type: string
          format: uuid
        proposalId:
          type: string
          format: uuid
        readyConfirmed:
          type: boolean
        readyDeadline:
          type: string
          format: date-time
        region:
          type: string
        rulesetId:
          type: string
          format: uuid
        rulesetVersion:
          type: integer
          minimum: 1
        status:
          type: string
          enum:
            - queued
            - proposed
            - matched
            - cancelled
        ticketId:
          type: string
          format: uuid
        updatedAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/ApiError"
    ApiError:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - traceId
      properties:
        code:
          type: string
          pattern: ^[a-z][a-z0-9_]*$
          maxLength: 64
          description: Stable machine-readable error code.
        message:
          type: string
          minLength: 1
          maxLength: 500
          description: Safe human-readable explanation.
        traceId:
          type: string
          minLength: 1
          maxLength: 128
          description: Identifier supplied when contacting support.
        details:
          type: object
          additionalProperties: true
          description: Safe structured context specific to the error code.
...
  responses:
    BadRequest:
      description: The request is malformed or fails validation.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication is missing, expired, or invalid.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
        WWW-Authenticate:
          $ref: "#/components/headers/WWWAuthenticate"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: The authenticated account cannot perform this operation.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: The request conflicts with current authoritative state.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: The requested account or profile does not exist.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServiceUnavailable:
      description: Identity services are unavailable or not configured.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"